Simplify slugs for DeFi apps
