+++ ensawards.org/data/apps/lido-defi/benchmarks/index.tsx
@@ -2,7 +2,7 @@
 // for additional advice on adding and modifying app benchmarks
 
 import type { AcceptanceTestBenchmark } from "data/acceptance-tests/types";
-import UniswapDeFiApp from "data/apps/uniswap-defi-app";
+import LidoDeFi from "data/apps/lido-defi";
 import { defineAppBenchmarks } from "data/benchmarks/registry";
 import { BenchmarkResults } from "data/benchmarks/types";
 import contributors from "data/contributors";
@@ -13,7 +13,8 @@ import { parseTimestamp } from "@ensnode/ensnode-sdk";
 
 import { cn } from "@/utils/tailwindClassConcatenation";
 
-import correctlyResolveEnsv2TestNameAddressProofImage from "./correctly-resolve-ensv2-test-name-address-proof.png";
+import correctlyResolveEnsv2TestNameAddressProofImageName from "./correctly-resolve-ensv2-test-name-address-proof.png";
+import correctlyResolveEnsv2TestNameAddressProofImageAddress from "./correctly-resolve-ensv2-test-name-address-proof-address.png";
 
 const benchmarks: BestPracticeBenchmarks = {
   // TODO: `Contract Naming` category is temporarily hidden due to unfit content,
@@ -28,27 +29,35 @@ const benchmarks: BestPracticeBenchmarks = {
   },
   "ensv2-ready-resolution": {
     "correctly-resolve-ensv2-test-name-address": {
-      result: BenchmarkResults.Pass,
+      result: BenchmarkResults.PartialPass,
       contributions: [
-        { from: contributors.y3drk, lastUpdated: parseTimestamp("2026-06-09T11:00:21Z") },
+        { from: contributors.y3drk, lastUpdated: parseTimestamp("2026-06-09T13:49:30Z") },
       ],
       notes: (
         <div className={cn(acceptanceTestDetailsContainerStyles, "w-full")}>
           <p className="w-full">
-            ENSv2 ready resolution was tested using the &quot;search&quot; tool. The resolved
-            address is correct.
+            ENSv2 ready resolution was tested using the &quot;rewards&quot; explorer. The app
+            doesn't explicitly prompt the user to pass an ENS name as identifier, but when it
+            receives one, the resolved address is correct.
           </p>
-          <img
-            alt="Uniswap DeFi app correctly resolves the name for ENSv2"
-            src={correctlyResolveEnsv2TestNameAddressProofImage.src}
-            className="w-auto h-full max-h-[325px] rounded-xl"
-          />
+          <div className="flex flex-col justify-start items-center gap-5">
+            <img
+              alt="Lido DeFi app correctly resolves the name for ENSv2"
+              src={correctlyResolveEnsv2TestNameAddressProofImageName.src}
+              className="w-auto h-full max-h-[325px] rounded-xl"
+            />
+            <img
+              alt="Proof that the resolution flow is the same for address and name inputs"
+              src={correctlyResolveEnsv2TestNameAddressProofImageAddress.src}
+              className="w-auto h-full max-h-[325px] rounded-xl"
+            />
+          </div>
         </div>
       ),
     } as const satisfies AcceptanceTestBenchmark,
   },
 } as const satisfies BestPracticeBenchmarks;
 
-defineAppBenchmarks(UniswapDeFiApp, benchmarks);
+defineAppBenchmarks(LidoDeFi, benchmarks);
 
 export default benchmarks;
