+++ ensawards.org/data/apps/uniswap-defi/benchmarks/index.tsx
@@ -1,0 +1,54 @@
+// Read https://github.com/namehash/ensawards/blob/main/CONTRIBUTING.md
+// for additional advice on adding and modifying app benchmarks
+
+import type { AcceptanceTestBenchmark } from "data/acceptance-tests/types";
+import UniswapDeFi from "data/apps/uniswap-defi";
+import { defineAppBenchmarks } from "data/benchmarks/registry";
+import { BenchmarkResults } from "data/benchmarks/types";
+import contributors from "data/contributors";
+import { acceptanceTestDetailsContainerStyles } from "data/ens-best-practices/styles";
+import type { BestPracticeBenchmarks } from "data/ens-best-practices/types.ts";
+
+import { parseTimestamp } from "@ensnode/ensnode-sdk";
+
+import { cn } from "@/utils/tailwindClassConcatenation";
+
+import correctlyResolveEnsv2TestNameAddressProofImage from "./correctly-resolve-ensv2-test-name-address-proof.png";
+
+const benchmarks: BestPracticeBenchmarks = {
+  // TODO: `Contract Naming` category is temporarily hidden due to unfit content,
+  // and so are all benchmarks belonging to it.
+  // We aim to fix it as soon as we have the capacity.
+  // See: https://github.com/namehash/ensawards/issues/222
+  "display-named-smart-contracts-mainnet": {
+    "mainnet-interactions-display-named-smart-contracts": undefined,
+  },
+  "display-named-smart-contracts-l2-chains": {
+    "l2-chain-interactions-display-named-smart-contracts": undefined,
+  },
+  "ensv2-ready-resolution": {
+    "correctly-resolve-ensv2-test-name-address": {
+      result: BenchmarkResults.Pass,
+      contributions: [
+        { from: contributors.y3drk, lastUpdated: parseTimestamp("2026-06-09T11:00:21Z") },
+      ],
+      notes: (
+        <div className={cn(acceptanceTestDetailsContainerStyles, "w-full")}>
+          <p className="w-full">
+            ENSv2 ready resolution was tested using the &quot;search&quot; tool. The resolved
+            address is correct.
+          </p>
+          <img
+            alt="Uniswap DeFi app correctly resolves the name for ENSv2"
+            src={correctlyResolveEnsv2TestNameAddressProofImage.src}
+            className="w-auto h-full max-h-[325px] rounded-xl"
+          />
+        </div>
+      ),
+    } as const satisfies AcceptanceTestBenchmark,
+  },
+} as const satisfies BestPracticeBenchmarks;
+
+defineAppBenchmarks(UniswapDeFi, benchmarks);
+
+export default benchmarks;
